Extend icons available in "social" sidebar block
The template adds "-square" to the Font Awesome icon to be used,
unless the icon name is on the exception list. Some icon names
potentially relevant for the "social" sidebar block do not have
a "-square" version but are not on the exception list, so the
icons do not display correctly.

Add more relevant icon names to the list: 'home', 'user', 'users',
'envelope', 'envelope-o', 'stack-exchange', 'hacker-news'.