Fix display of "About Me" in pelican-bootstrap3
The "About Me" block (defined in templates/includes/aboutme.html)
consists of the image defined by AVATAR and the text in ABOUT_ME.
However, the template only displays the block if ABOUT_ME is defined.
If only AVATAR is defined, then the block is not displayed at all.

- Change the test so the "About Me" block is displayed if either
ABOUT_ME or AVATAR is defined.

- Change the structure of the "About Me" block so that the heading for
the ABOUT_ME text is displayed only if ABOUT_ME text is defined,
to avoid an empty heading. If AVATAR is defined but ABOUT_ME is not,
the block displays the AVATAR image only.