Adds DNT meta so that Twitter doesn't track us
As explained on this page, this meta tag prevents Twitter from using your webpage to track people:

https://dev.twitter.com/web/overview/privacy#what-privacy-options-do-website-publishers-have

Generally speaking, this is  the kind of tracking that nobody expects. Nobody thinks that by going to *my* website, which shows some tweets, that Twitter will be tracking them.

Let's turn that off by default. Hopefully this isn't controversial.