:sparkles: Add CC Zero to the cc_license_mark macro
In getpelican/pelican-themes#530 is reported that the cc_license_mark macro, written in templates/includes/cc-license.html, doesn't get the correct url and image when CC_LICENSE is set to "zero" or "cc0". That's because the format of the url is different. This change add an if to know if it is a CC Public Domain (Zero) mark before to set the license.